Make it compile on mac (mac types are slightly other defined)
git-svn-id: svn://maxr.org/MAX/trunk@1606 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ clist.h
@@ -42,7 +42,7 @@ template<typename T> class cList
 
 template<typename T> void cList<T>::Add(T const& e)
 {
-	if (size_ >= capacity_) Reserve(std::max(1U, size_ * 2));
+	if (size_ >= capacity_) Reserve(std::max((size_t)1U, size_ * 2));
 	new (&v_[size_]) T(e);
 	++size_;
 }
@@ -50,7 +50,7 @@ template<typename T> void cList<T>::Add(T const& e)
 template<typename T> void cList<T>::Insert( size_t const i, T const& e )
 {
 	if ( i > size_ ) throw;
-	if (size_ >= capacity_) Reserve(std::max(1U, size_ * 2));
+	if (size_ >= capacity_) Reserve(std::max((size_t)1U, size_ * 2));
 
 	for (size_t n = size_; n > i; --n ) v_[n] = v_[n - 1];
 	new (&v_[i]) T(e);
